Started moving most modal behaviour into smaller files
