+++ js/modals/question-modal/events/input-output/remove-pair.js
@@ -1,0 +1,12 @@
+function removeInputOutputPair() {
+   'use strict';
+
+   var $inputOutputButtons = $('.modal .js-input-output-button');
+
+   if ($inputOutputButtons.length > 3) {
+      $inputOutputButtons.last().remove();
+   }
+}
+
+module.exports = removeInputOutputPair;
+
