+++ js/modals/question-modal/events/input-output/save-input.js
@@ -1,0 +1,8 @@
+
+function saveInput(inputText) {
+   'use strict';
+
+   $('.js-input-output-active').data('input', inputText);
+}
+
+module.exports = saveInput;
