+++ js/modals/question-modal/events/input-output/save-output.js
@@ -1,0 +1,8 @@
+
+function saveOutput(outputText) {
+   'use strict';
+
+   $('.js-input-output-active').data('output', outputText);
+}
+
+module.exports = saveOutput;
